fix: refactor splitBatchIntoTwo, add related testcases
